move Pebble project schemas to Pebble module
